DevMaster Debugger/ai-debug-companion/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-debug-companion"
version = "0.1.0"
description = "An intelligent TUI tool that watches your development session and suggests fixes"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "AI Debug Team"}
]
dependencies = [
    "textual>=0.47.0",
    "GitPython>=3.1.40",
    "watchdog>=3.0.0",
    "rich>=13.7.0",
    "click>=8.1.7",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.12.0",
    "ruff>=0.1.8",
]

[project.scripts]
debug-companion = "ai_debug_companion.cli:main"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
